+++ src/components/AuthBlock/AuthButton/AuthButton.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .button {
   background-color: transparent;
