+++ src/components/AuthBlock/AuthInput/AuthInput.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .form {
   &__item {
