+++ src/components/AuthBlock/AuthInput/AuthInput.tsx
