+++ src/components/AuthBlock/SignUpController.tsx
@@ -1,30 +1,39 @@
-import { useContext, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import Router from 'next/router';
+import { useRouter } from 'next/router';
 import Link from 'next/link';
 import { schema } from '@/validation/validationSchema';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { getAuthError } from '@/utils/getAuthError';
+import { useAuth } from '@/context/AuthProvider';
 import { ROUTES } from '@/constants/routes';
-import AuthInput from '../AuthInput/AuthInput';
-import AuthButton from '../AuthButton/AuthButton';
+import AuthButton from './AuthButton/AuthButton';
+import AuthInput from './AuthInput/AuthInput';
 import { AuthViewProps, LangContext, schemaType } from '@/types/types';
-
 import styles from './style.module.scss';
 import langContext from '@/context/langContext';
 import langChecker from '@/utils/langChecker';
 
-const SignInController = ({ authCallback }: AuthViewProps) => {
+const SignUpController = ({ authCallback }: AuthViewProps) => {
   const [isVisible, setIsVisible] = useState(false);
   const [authError, setAuthError] = useState<string>('');
   const [loading, setLoading] = useState(false);
-  const context: LangContext = useContext<LangContext>(langContext);
+  const context = useContext<LangContext>(langContext);
+  const Router = useRouter();
+  const { user } = useAuth();
 
   const checkedLang = langChecker(Router, context);
 
   const handlePasswordVisibility = (): void => {
     setIsVisible(!isVisible);
   };
+
+  useEffect(() => {
+    if (user) {
+      setLoading(false);
+      Router.push(ROUTES.MAIN);
+    }
+  }, [user, Router]);
 
   const {
     register,
@@ -36,7 +45,6 @@ const SignInController = ({ authCallback }: AuthViewProps) => {
     setLoading(true);
     try {
       await authCallback(email, password);
-      await Router.push(ROUTES.MAIN);
     } catch (e) {
       const err = getAuthError(e);
       setAuthError(err);
@@ -50,21 +58,21 @@ const SignInController = ({ authCallback }: AuthViewProps) => {
         <form className={styles['form']} onSubmit={onSubmit}>
           <div className={styles['form__title-container']}>
             <h3 className={styles['form__title']}>
-              {context.getConstants().signIn}
+              {context.getConstants().signUpTitle}
             </h3>
             {authError && (
               <p className={styles['form__error']}>
-                {context.getConstants().incorrect}
+                {context.getConstants().alreadyExists}
               </p>
             )}
           </div>
           <p className={styles['form__account']}>
-            {context.getConstants().haveAccount}{' '}
+            {context.getConstants().alreadyRegistered}{' '}
             <Link
               className={styles['form__link']}
-              href={ROUTES.SIGN_UP + `?lang=${checkedLang}`}
+              href={ROUTES.SIGN_IN + `?lang=${checkedLang}`}
             >
-              {context.getConstants().signUpHere}
+              {context.getConstants().signInHere}
             </Link>
             <br />
             {context.getConstants().signOr}{' '}
@@ -97,8 +105,9 @@ const SignInController = ({ authCallback }: AuthViewProps) => {
               handlePasswordVisibility={handlePasswordVisibility}
             />
           </div>
+
           <AuthButton
-            text={context.getConstants().signIn}
+            text={context.getConstants().signUp}
             type="submit"
             isLoading={loading}
             isDisabled={!isValid || loading}
@@ -109,4 +118,4 @@ const SignInController = ({ authCallback }: AuthViewProps) => {
   );
 };
 
-export default SignInController;
+export default SignUpController;
