+++ src/components/AuthBlock/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .form {
   background-color: $background-color;
