+++ src/components/Editor/Editor.module.scss
@@ -1,4 +1,4 @@
-@import './../../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .editor {
   background-color: rgba(0, 0, 0, 0);
