+++ src/components/Editor/Editor.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import Image from 'next/image';
 import { useState } from 'react';
 import CodeMirror from '@uiw/react-codemirror';
-import useGetWindowDimensions from '../../utils/useGetWindowsDimensions';
-import Tabs from '../Tabs/Tabs';
-import Button from '../Button/Button';
+import useGetWindowDimensions from '@/utils/useGetWindowsDimensions';
+import Tabs from '@/components/Tabs/Tabs';
+import Button from '@/components/Button/Button';
 import { TEditor } from '@/types/types';
 import { prettify } from '@/utils/prettify';
 import { CLEAN_IMAGE, PLAY_IMAGE } from '@/constants/buttonsImages';
