+++ src/components/Header/Header.module.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@import '../../styles/mixins';
-@import '../../styles/colors';
+@import '@/styles/mixins';
+@import '@/styles/colors';
 
 .header {
   @include animateHeader;
@@ -33,7 +33,6 @@
     display: flex;
     width: 320px;
     height: 80px;
-    margin-right: 20px;
     justify-content: space-evenly;
     align-items: center;
