+++ src/components/LangButton/LangButton.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors';
+@import '@/styles/colors';
 
 .switch {
   display: inline-block;
