+++ src/components/LangButton/LangButton.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect, useState } from 'react';
-import { LangContext } from '../../types/types';
-import LanguageContext from '../../context/langContext';
+import { LangContext } from '@/types/types';
+import LanguageContext from '@/context/langContext';
 import { NextRouter, useRouter } from 'next/router';
 import styles from './LangButton.module.scss';
