+++ src/components/NotFound/NotFound.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .wrapper {
   align-items: center;
