+++ src/styles/Home.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .main {
   display: flex;
