+++ src/styles/mixins.scss
@@ -96,7 +96,7 @@
   box-sizing: border-box;
   height: $heightParam;
   margin: 0 auto;
-  padding: 5px 15px;
+  padding: 5px 25px;
   justify-content: space-between;
   align-items: center;
   background: $background;
